UI: basic refactoring (just code movement, really)
