+++ Source/UI/UIHolder.cs
@@ -1,0 +1,79 @@
+﻿using System;
+using UnityEngine;
+using KSP.UI.Screens;
+
+namespace RP0
+{
+    [KSPAddon(KSPAddon.Startup.SpaceCentre, false)]
+    class UIHolder : MonoBehaviour
+    {
+        // GUI
+        static bool guiEnabled = false;
+        private ApplicationLauncherButton button;
+        private TopWindow tw = new TopWindow();
+
+        protected void Awake()
+        {
+            try {
+                GameEvents.onGUIApplicationLauncherReady.Add(this.OnGuiAppLauncherReady);
+            } catch (Exception ex) {
+                Debug.LogError("RP0 failed to register UIHolder.OnGuiAppLauncherReady");
+                Debug.LogException(ex);
+            }
+        }
+
+        private void ShowWindow()
+        {
+            guiEnabled = true;
+        }
+        private void HideWindow()
+        {
+            guiEnabled = false;
+        }
+
+        private void OnSceneChange(GameScenes s)
+        {
+            if (s != GameScenes.SPACECENTER)
+                HideWindow();
+        }
+
+        private void OnGuiAppLauncherReady()
+        {
+            if (HighLogic.CurrentGame.Mode != global::Game.Modes.CAREER)
+                return;
+            try {
+                button = ApplicationLauncher.Instance.AddModApplication(
+                    ShowWindow,
+                    HideWindow,
+                    null,
+                    null,
+                    null,
+                    null,
+                    ApplicationLauncher.AppScenes.SPACECENTER,
+                    GameDatabase.Instance.GetTexture("RP-0/maintecost", false));
+                GameEvents.onGameSceneLoadRequested.Add(this.OnSceneChange);
+            } catch (Exception ex) {
+                Debug.LogError("RP0 failed to register UIHolder");
+                Debug.LogException(ex);
+            }
+        }
+
+        public void OnDestroy()
+        {
+            try {
+                GameEvents.onGUIApplicationLauncherReady.Remove(this.OnGuiAppLauncherReady);
+                if (button != null)
+                    ApplicationLauncher.Instance.RemoveModApplication(button);
+            } catch (Exception ex) {
+                Debug.LogException(ex);
+            }
+        }
+
+        public void OnGUI()
+        {
+            if (guiEnabled)
+                tw.OnGUI();
+        }
+    }
+}
+
